Moved code related to modal models into separate files
