+++ lib/pythontarget.c
@@ -33,6 +33,7 @@ THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include "pythontarget.h"
 #include "core/utils/util.h"
 #include "core/tag.h"
+#include "modal_models/definitions.h"
 
 PyTypeObject TagType;
 
@@ -94,30 +95,6 @@ static PyObject* py_port_set(PyObject *self, PyObject *args) {
         p->value = val;
         p->is_present = true;
     }
-
-    Py_INCREF(Py_None);
-    return Py_None;
-}
-
-/**
- * Set a new mode for a modal model.
- */
-static PyObject* py_mode_set(PyObject *mode_capsule, PyObject *args) {
-    mode_capsule_struct_t* m = (mode_capsule_struct_t*)mode_capsule;
-
-    reactor_mode_t* mode = PyCapsule_GetPointer(m->mode, "mode");
-    if (mode == NULL) {
-        error_print("Null pointer received.");
-        exit(1);
-    }
-
-    self_base_t* self = PyCapsule_GetPointer(m->lf_self, "lf_self");
-    if (self == NULL) {
-        error_print("Null pointer received.");
-        exit(1);
-    }
-
-    _LF_SET_MODE_WITH_TYPE(mode, m->change_type);
 
     Py_INCREF(Py_None);
     return Py_None;
@@ -828,32 +805,6 @@ static PyObject *Tag_richcompare(py_tag_t *self, PyObject *other, int op) {
 //////////////////////////////////////////////////////////////
 /////////////  Python Structs
 
-////// Modes //////
-
-/*
- * The function members of mode_capsule.
- * The set function is used to set a new mode.
- */
-static PyMethodDef mode_capsule_methods[] = {
-    {"set", (PyCFunction)py_mode_set, METH_NOARGS, "Set a new mode."},
-    {NULL}  /* Sentinel */
-};
-
-/*
- * The definition of mode_capsule type object, which is
- * used to describe how mode_capsule behaves.
- */
-static PyTypeObject mode_capsule_t = {
-    PyVarObject_HEAD_INIT(NULL, 0)
-    .tp_name = "LinguaFranca.mode_capsule",
-    .tp_doc = "mode_capsule objects",
-    .tp_basicsize = sizeof(mode_capsule_struct_t),
-    .tp_itemsize = 0,
-    .tp_flags = Py_TPFLAGS_DEFAULT,
-    .tp_new = PyType_GenericNew,
-    .tp_methods = mode_capsule_methods,
-};
-
 ////// Ports //////
 /*
  * The members of a port_capsule, used to define
@@ -1078,11 +1029,6 @@ PyMODINIT_FUNC
 GEN_NAME(PyInit_,MODULE_NAME)(void) {
     PyObject *m;
 
-    // Initialize the mode_capsule type
-    if (PyType_Ready(&mode_capsule_t) < 0) {
-        return NULL;
-    }
-
     // Initialize the port_capsule type
     if (PyType_Ready(&port_capsule_t) < 0) {
         return NULL;
@@ -1109,13 +1055,7 @@ GEN_NAME(PyInit_,MODULE_NAME)(void) {
         return NULL;
     }
 
-    // Add the mode_capsule type to the module's dictionary.
-    Py_INCREF(&mode_capsule_t);
-    if (PyModule_AddObject(m, "mode_capsule", (PyObject *) &mode_capsule_t) < 0) {
-        Py_DECREF(&mode_capsule_t);
-        Py_DECREF(m);
-        return NULL;
-    }
+    initialize_mode_capsule_t(m);
 
     // Add the port_capsule type to the module's dictionary
     Py_INCREF(&port_capsule_t);
@@ -1152,8 +1092,6 @@ GEN_NAME(PyInit_,MODULE_NAME)(void) {
     return m;
 }
 
-
-
 //////////////////////////////////////////////////////////////
 /////////////  Python Helper Functions
 /// These functions are called in generated C code for various reasons.
@@ -1164,41 +1102,6 @@ GEN_NAME(PyInit_,MODULE_NAME)(void) {
  **/
 void destroy_action_capsule(PyObject* capsule) {
     free(PyCapsule_GetPointer(capsule, "action"));
-}
-
-/**
- * Convert a `reactor_mode_t` to a `mode_capsule_t`.
- */
-PyObject* convert_C_mode_to_py(
-		reactor_mode_t* mode,
-		self_base_t* lf_self,
-		lf_mode_change_type_t change_type
-) {
-    // Create the mode struct in Python
-	mode_capsule_struct_t* cap =
-        (mode_capsule_struct_t*)PyObject_GC_New(mode_capsule_struct_t, &mode_capsule_t);
-    if (cap == NULL) {
-        error_print_and_exit("Failed to convert mode.");
-    }
-
-    // Create the capsule to hold the reactor_mode_t* mode
-    PyObject* capsule = PyCapsule_New(mode, "mode", NULL);
-    if (capsule == NULL) {
-        error_print_and_exit("Failed to convert mode.");
-    }
-    // Fill in the Python mode struct.
-    cap->mode = capsule;
-
-    // Create a capsule to point to the self struct.
-    PyObject* self_capsule = PyCapsule_New(lf_self, "lf_self", NULL);
-    if (self_capsule == NULL) {
-        error_print_and_exit("Failed to convert self.");
-    }
-    cap->lf_self = self_capsule;
-
-    cap->change_type = change_type;
-
-	return cap;
 }
 
 /**
